merge

+++ pkg/utils/date-time_test.go
@@ -75,7 +75,7 @@ func TestGetDaysFromYearMonth(t *testing.T) {
 
 func TestGetWeekDayByYearMonth(t *testing.T) {
 	for i := 0; i < 12; i++ {
-		yearMonth := 201901 + i
+		yearMonth := int64(201901 + i)
 		month, _ := GetWeekDayByYearMonth(yearMonth)
 		for i, t := range month {
 			fmt.Println(i, t)
